I did some simple cleanup and added the about us page html from issue 26 (#28)
* Reference #22 - Reports that the imported inactive orgs are being duplicate.

The issue is #comment when the pull:orgs command was ran when checking for duplicate orgs I was ignoring the ones that are already deactivated.

I also added an option on the pull:orgs command to remove duplicates. It looks first the 2nd instance of the first inactive org and deletes from that id on. to run it

`php artisan pull:orgs --org-cleanup`

* I did some simple cleanup and added the about us page html from issue 26

* removed the link for @webdood and removed nofollow and _blank link attributes